+++ index.html
@@ -54,6 +54,6 @@
 </footer>
 
 
-    <script src="./modules/index.js" type="module"></script>
+    <script src="./lib/index.js" type="module"></script>
 </body>
 </html>
