+++ lib/constructor.js
@@ -1,4 +1,3 @@
-
 export default class BookList {
   constructor() {
     this.bookList = document.getElementById('book-list');
