+++ package.json
@@ -16,5 +16,18 @@
   "bugs": {
     "url": "https://github.com/nilab-khwaja/awesome_books_ES6/issues"
   },
-  "homepage": "https://github.com/nilab-khwaja/awesome_books_ES6#readme"
+  "homepage": "https://github.com/nilab-khwaja/awesome_books_ES6#readme",
+  "devDependencies": {
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.27.5",
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0"
+  },
+  "dependencies": {
+    "luxon": "^3.3.0"
+  }
 }
