+++ styles.css
@@ -1,99 +1,99 @@
 * {
-    box-sizing: border-box;
-  }
-  
-  body {
-    padding: 10px;
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-    justify-content: stretch;
-  }
-  
-  .container {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    gap: 4px;
-    flex-grow: 1;
-  }
-  
-  nav {
-    display: flex;
-    width: 100%;
-    flex-direction: row;
-    align-items: baseline;
-    gap: 1rem;
-    border: solid 3px black;
-    padding: 0.5rem;
-  }
-  
-  nav * {
-    font-size: 16px;
-    text-decoration: none;
-    color: black;
-    margin: 0;
-  }
-  
-  h1 {
-    flex-grow: 1;
-  }
-  
-  .collection {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 4px;
-  }
-  
-  #list-section {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 50%;
-  }
-  
-  #book-list {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    border: solid 3px black;
-  }
-  
-  .collection:nth-child(odd) {
-    background-color: gray;
-  }
-  
-  .add-book {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1rem;
-  }
-  
-  .add-book input {
-    width: 600px;
-    padding: 10px;
-    border: 1px solid black;
-    box-shadow: 2px 2px 2px;
-  }
-  
-  .add-book button {
-    align-self: flex-end;
-  }
-  
-  #add-section {
-    display: none;
-  }
-  
-  #contact-section {
-    display: none;
-    flex-direction: column;
-    align-items: center;
-  }
-  
-  footer {
-    padding: 2rem;
-  }
+  box-sizing: border-box;
+}
+
+body {
+  padding: 10px;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+  justify-content: stretch;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  gap: 4px;
+  flex-grow: 1;
+}
+
+nav {
+  display: flex;
+  width: 100%;
+  flex-direction: row;
+  align-items: baseline;
+  gap: 1rem;
+  border: solid 3px black;
+  padding: 0.5rem;
+}
+
+nav * {
+  font-size: 16px;
+  text-decoration: none;
+  color: black;
+  margin: 0;
+}
+
+h1 {
+  flex-grow: 1;
+}
+
+.collection {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 4px;
+}
+
+#list-section {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 50%;
+}
+
+#book-list {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  border: solid 3px black;
+}
+
+.collection:nth-child(odd) {
+  background-color: gray;
+}
+
+.add-book {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+}
+
+.add-book input {
+  width: 600px;
+  padding: 10px;
+  border: 1px solid black;
+  box-shadow: 2px 2px 2px;
+}
+
+.add-book button {
+  align-self: flex-end;
+}
+
+#add-section {
+  display: none;
+}
+
+#contact-section {
+  display: none;
+  flex-direction: column;
+  align-items: center;
+}
+
+footer {
+  padding: 2rem;
+}
